feat(parametermanager): Added get samples for parametermanager
